Cell file IO reimplemented

+++ modules/chain/include/dap_chain_cell.h
@@ -78,5 +78,5 @@ dap_chain_cell_t *dap_chain_cell_find_by_id(dap_chain_t *a_chain, dap_chain_cell
 void dap_chain_cell_close(dap_chain_cell_t *a_cell);
 void dap_chain_cell_delete(dap_chain_cell_t *a_cell);
 int dap_chain_cell_load(dap_chain_t *a_chain, dap_chain_cell_t *a_cell);
-int dap_chain_cell_file_update(dap_chain_cell_t *a_cell);
-int dap_chain_cell_file_append(dap_chain_cell_t *a_cell,const void *a_atom, size_t a_atom_size);
+ssize_t dap_chain_cell_file_update(dap_chain_cell_t *a_cell);
+ssize_t dap_chain_cell_file_append(dap_chain_cell_t *a_cell,const void *a_atom, size_t a_atom_size);
